Checks for usage of facility group type before deleting the type.

+++ apps/frontend/modules/scenario/actions/actions.class.php
@@ -1431,9 +1431,26 @@ class scenarioActions extends agActions
     $this->forward404Unless($ag_facility_group_type = Doctrine_Core::getTable('agFacilityGroupType')->find(array($request->getParameter('id'))),
                                                                                                                                         sprintf('Object ag_facility_group_type does not exist (%s).',
                                                                                                                                                 $request->getParameter('id')));
-    $ag_facility_group_type->delete();
+    $facilityGroupTypeId = $request->getParameter('id');
+    $scenarioFacilityGroupByTypeCount = agDoctrineQuery::create()
+                                        ->select('COUNT(g.id)')
+                                        ->from('agScenarioFacilityGroup AS g')
+                                        ->where('g.facility_group_type_id = ?', $facilityGroupTypeId)
+                                        ->execute(array(), Doctrine_Core::HYDRATE_SINGLE_SCALAR);
 
-    $this->redirect('scenario/grouptype');
+    $eventFacilityGroupByTypeCount = agDoctrineQuery::create()
+                                     ->select('COUNT(g.id)')
+                                     ->from('agEventFacilityGroup AS g')
+                                     ->where('g.facility_group_type_id = ?', $facilityGroupTypeId)
+                                     ->execute(array(), Doctrine_Core::HYDRATE_SINGLE_SCALAR);
+
+    if ( ($scenarioFacilityGroupByTypeCount == 0) && ($eventFacilityGroupByTypeCount == 0) )
+    {
+      $ag_facility_group_type->delete();
+
+      $this->redirect('scenario/grouptype');
+    }
+    $this->facilityGroupType = Doctrine_Core::getTable('agFacilityGroupType')->find($facilityGroupTypeId)->getFacilityGroupType();
   }
 
   /**
